Investigating chirping signals
Here is a code that investigates the WDM transform on a chirping sinusoid.
It was originally used to debug a few problems with backward transform that
appears to now be fixed.

This code simply plots the wavelet domain representation of a chirp and checks
the backwards transformation. It can be used as a tutorial in essence