Likelihood built for gated noise in wavelet domain
Building likelihood across amplitude values for quasi-monochromatic signal. With noise
I observe reasonable statistical fluctuations regardless of what choice of Nf I make.

I have to process the data after it is generated though. I take a template, split it into chunks
and combine it in the same way as the input data. This indicates that the original signal generation
(with masks for gaps) is not consistent. Work on this tomorrow.